python/requirements.txt
pika
pyrabbit
python-dotenv
